EVSRESTPAPI-449 reorganized the fhir classes, updated imports
