+++ src/main/java/gov/nih/nci/evs/api/fhir/R4/CodeSystemProviderR4.java
@@ -15,8 +15,8 @@ import ca.uhn.fhir.rest.param.TokenParam;
 import ca.uhn.fhir.rest.server.IResourceProvider;
 import ca.uhn.fhir.rest.server.servlet.ServletRequestDetails;
 import gov.nih.nci.evs.api.controller.ConceptController;
-import gov.nih.nci.evs.api.fhir.FHIRServerResponseException;
-import gov.nih.nci.evs.api.fhir.FhirUtility;
+import gov.nih.nci.evs.api.util.FHIRServerResponseException;
+import gov.nih.nci.evs.api.util.FhirUtility;
 import gov.nih.nci.evs.api.model.Concept;
 import gov.nih.nci.evs.api.model.IncludeParam;
 import gov.nih.nci.evs.api.model.Terminology;
