+++ src/main/java/gov/nih/nci/evs/api/fhir/R4/ConceptMapProviderR4.java
@@ -13,8 +13,8 @@ import ca.uhn.fhir.rest.param.StringParam;
 import ca.uhn.fhir.rest.param.TokenParam;
 import ca.uhn.fhir.rest.server.IResourceProvider;
 import ca.uhn.fhir.rest.server.servlet.ServletRequestDetails;
-import gov.nih.nci.evs.api.fhir.FHIRServerResponseException;
-import gov.nih.nci.evs.api.fhir.FhirUtility;
+import gov.nih.nci.evs.api.util.FHIRServerResponseException;
+import gov.nih.nci.evs.api.util.FhirUtility;
 import gov.nih.nci.evs.api.model.Concept;
 import gov.nih.nci.evs.api.model.IncludeParam;
 import gov.nih.nci.evs.api.model.Property;
