+++ src/main/java/gov/nih/nci/evs/api/fhir/R4/package-info.java
@@ -1,0 +1,2 @@
+/** FHIR implementation. */
+package gov.nih.nci.evs.api.fhir.R4;
