--- src/main/java/gov/nih/nci/evs/api/fhir/package-info.java
@@ -1,2 +1,0 @@
-/** FHIR implementation. */
-package gov.nih.nci.evs.api.fhir;
