+++ src/main/java/gov/nih/nci/evs/api/util/FHIRServerResponseException.java
@@ -1,4 +1,4 @@
-package gov.nih.nci.evs.api.fhir;
+package gov.nih.nci.evs.api.util;
 
 import ca.uhn.fhir.rest.server.exceptions.BaseServerResponseException;
 import org.hl7.fhir.instance.model.api.IBaseOperationOutcome;
