+++ src/main/java/gov/nih/nci/evs/api/util/FhirUtility.java
@@ -7,7 +7,7 @@
  * and are protected by trade secret or copyright law.  Dissemination of this information
  * or reproduction of this material is strictly forbidden.
  */
-package gov.nih.nci.evs.api.fhir;
+package gov.nih.nci.evs.api.util;
 
 import static java.lang.String.format;
 
@@ -15,6 +15,7 @@ import ca.uhn.fhir.rest.param.DateParam;
 import ca.uhn.fhir.rest.param.DateRangeParam;
 import ca.uhn.fhir.rest.param.StringParam;
 import java.util.Date;
+
 import org.apache.commons.lang3.StringUtils;
 import org.hl7.fhir.r5.model.CodeType;
 import org.hl7.fhir.r5.model.Coding;
